support void input products in more cases
